Trying to encrypt messages and decrypt them

+++ app/src/main/java/com/example/ofir/social_geha/Activities_and_Fragments/ChatActivity.java
@@ -132,11 +132,11 @@ public class ChatActivity extends AppCompatActivity {
 
         mFirestore = FirebaseFirestore.getInstance();
 
-        setMessageListners();
+        setMessageListeners();
     }
 
 
-    private void setMessageListners() {
+    private void setMessageListeners() {
         mFirestore.collection(MESSAGES).whereEqualTo("toPersonID", mLoggedInPersonId)
                 .whereEqualTo("fromPersonID", mOtherPersonId)
                 .addSnapshotListener(new EventListener<QuerySnapshot>() {
@@ -147,12 +147,15 @@ public class ChatActivity extends AppCompatActivity {
                             return;
                         }
 
+                        if (aes == null)
+                            throw new AssertionError("AES should not be null");
                         for (DocumentChange doc : queryDocumentSnapshots.getDocumentChanges()) {
                             if (doc.getType() == DocumentChange.Type.ADDED) {
                                 String message_text = doc.getDocument().getString("message");
                                 Log.d("COOLTEST", "Content: " + message_text);
                                 Log.d("POPO", "onEvent: DOES THIS");
                                 Message message = doc.getDocument().toObject(Message.class);
+                                message.setMessage(aes.decrypt(message.getMessage().getBytes()));
                                 mFirestore.collection(MESSAGES).document(doc.getDocument().getId()).delete();
                                 if (message.getShown())
                                     messageList.add(message);
@@ -187,7 +190,8 @@ public class ChatActivity extends AppCompatActivity {
             Database.getInstance().sendControlMessage("AES" + AES.keyToString(aes.key), mLoggedInPersonId, mOtherPersonId);
             Log.d("AES_ENCRYPT_SEND", "Sending " + mOtherPersonId + " message:" + "AES" + AES.keyToString(aes.key));
         }
-        Database.getInstance().sendMessage(message, mLoggedInPersonId, mOtherPersonId);
+        Database.getInstance().sendMessage(String.valueOf(aes.encrypt(message)), mLoggedInPersonId, mOtherPersonId);
+        Log.d("AES_SEND_MESSAGE", String.valueOf(aes.encrypt(message)));
         Message mymessage = new Message(message, mLoggedInPersonId, mOtherPersonId, true);
         fileHandler.writeMessage(mymessage);
         messageList.add(mymessage);
